consolidate and fix imports for DCR
